+++ tweetshake/LoginViewController.swift
@@ -16,6 +16,8 @@ class LoginViewController: UIViewController {
     var credentialSetupMessageLabel: UILabel?
     var isLoginViewVisible = false
     var loginViewGeneration = 0
+    var loginAttemptGeneration = 0
+    var activeLoginAttemptGeneration: Int?
 
     override func viewWillAppear(animated: Bool) {
         super.viewWillAppear(animated)
@@ -30,6 +32,7 @@ class LoginViewController: UIViewController {
     override func viewWillDisappear(animated: Bool) {
         super.viewWillDisappear(animated)
         isLoginViewVisible = false
+        activeLoginAttemptGeneration = nil
         logInButton?.removeFromSuperview()
         logInButton = nil
     }
@@ -46,18 +49,24 @@ class LoginViewController: UIViewController {
 
     func installLoginButtonForCurrentAppearance() {
         logInButton?.removeFromSuperview()
-        let generation = loginViewGeneration
+        loginAttemptGeneration += 1
+        let appearanceGeneration = loginViewGeneration
+        let attemptGeneration = loginAttemptGeneration
+        activeLoginAttemptGeneration = attemptGeneration
         let logInButton = TWTRLogInButton(logInCompletion: { [weak self] (session: TWTRSession!, error: NSError!) in
             dispatch_async(dispatch_get_main_queue()) {
                 if let viewController = self {
-                    guard viewController.canHandleLoginCompletion(generation) else {
+                    guard viewController.reserveLoginCompletion(
+                        appearanceGeneration,
+                        attemptGeneration: attemptGeneration
+                    ) else {
                         return
                     }
 
                     if session != nil && error == nil {
                         viewController.performSegueWithIdentifier("shake", sender: viewController)
                     } else {
-                        viewController.showLoginRequiredMessage()
+                        viewController.showLoginRequiredMessage(appearanceGeneration)
                     }
                 }
             }
@@ -67,8 +76,27 @@ class LoginViewController: UIViewController {
         centerLoginButton()
     }
 
-    func canHandleLoginCompletion(generation: Int) -> Bool {
-        return isLoginViewVisible && generation == loginViewGeneration
+    func reserveLoginCompletion(appearanceGeneration: Int, attemptGeneration: Int) -> Bool {
+        guard isLoginViewVisible &&
+              appearanceGeneration == loginViewGeneration &&
+              activeLoginAttemptGeneration == attemptGeneration else {
+            return false
+        }
+
+        activeLoginAttemptGeneration = nil
+        logInButton?.removeFromSuperview()
+        logInButton = nil
+        return true
+    }
+
+    func canRetryLogin(appearanceGeneration: Int) -> Bool {
+        return isLoginViewVisible && appearanceGeneration == loginViewGeneration
+    }
+
+    func restoreLoginAfterFailure(appearanceGeneration: Int) {
+        if canRetryLogin(appearanceGeneration) {
+            installLoginButtonForCurrentAppearance()
+        }
     }
 
     override func viewDidLayoutSubviews() {
@@ -108,13 +136,16 @@ class LoginViewController: UIViewController {
         layoutCredentialSetupMessage()
     }
 
-    func showLoginRequiredMessage() {
+    func showLoginRequiredMessage(appearanceGeneration: Int) {
         if self.presentedViewController != nil {
+            restoreLoginAfterFailure(appearanceGeneration)
             return
         }
 
         let alert = UIAlertController(title: "Twitter Login Required", message: "Sign in with Twitter before composing a tweet.", preferredStyle: UIAlertControllerStyle.Alert)
-        let action = UIAlertAction(title: "OK", style: UIAlertActionStyle.Default, handler: nil)
+        let action = UIAlertAction(title: "OK", style: UIAlertActionStyle.Default, handler: { [weak self] _ in
+            self?.restoreLoginAfterFailure(appearanceGeneration)
+        })
         alert.addAction(action)
         self.presentViewController(alert, animated: true, completion: nil)
     }